Updated and published the SharpPipes package
SharpPipes has been released and pushed up to the nuget server.
