+++ SharpWrapper - Extensions/SharpPipes/Properties/AssemblyInfo.cs
@@ -19,7 +19,7 @@ using System.Resources;
 [assembly: Guid("73552094-035A-4DFB-BEC4-5116EC7993EF")]
 
 // Version information
-[assembly: AssemblyVersion("1.0.0.4")]
-[assembly: AssemblyFileVersion("1.0.0.4")]
+[assembly: AssemblyVersion("0.1.2.8")]
+[assembly: AssemblyFileVersion("0.1.2.8")]
 [assembly: NeutralResourcesLanguageAttribute( "en-US" )]
