+++ SharpWrapper - Extensions/SharpPipes/SharpPipes.csproj
@@ -7,8 +7,8 @@
     <ProjectGuid>{E43E4328-FFC8-499C-B1F1-19160101E9D3}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>SharpPassThruPipes</RootNamespace>
-    <AssemblyName>SharpPassThruPipes</AssemblyName>
+    <RootNamespace>SharpPipes</RootNamespace>
+    <AssemblyName>SharpPipes</AssemblyName>
     <TargetFrameworkVersion>v4.7.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <Deterministic>true</Deterministic>
@@ -43,11 +43,10 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="J2534Pipes\PassThruPipe.cs" />
+    <Compile Include="PassThruPipe.cs" />
     <Compile Include="PassThruPipeEnums.cs" />
-    <Compile Include="PassThruPipeEvents.cs" />
-    <Compile Include="J2534Pipes\PassThruPipeReader.cs" />
-    <Compile Include="J2534Pipes\PassThruPipeWriter.cs" />
+    <Compile Include="PassThruPipeReader.cs" />
+    <Compile Include="PassThruPipeWriter.cs" />
     <Compile Include="Properties\AssemblyInfo.cs">
       <AutoGen>True</AutoGen>
       <DesignTime>True</DesignTime>
@@ -76,6 +75,7 @@
   <ItemGroup>
     <Service Include="{508349B6-6B84-4DF5-91F0-309BEEBAD82D}" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <PropertyGroup>
     <PostBuildEvent>copy /Y "$(ProjectDir)\FulcrumShim_$(ConfigurationName).dll" "$(TargetDir)\FulcrumShim.dll"
